+++ augur/__init__.py
@@ -3,7 +3,6 @@ The top-level augur command which dispatches to subcommands.
 """
 
 import argparse
-import re
 import os
 import sys
 import importlib
@@ -13,7 +12,7 @@ from types import SimpleNamespace
 
 from .errors import AugurError
 from .io import print_err
-from .utils import first_line
+from .argparse_ import add_command_subparsers, add_default_command
 
 recursion_limit = os.environ.get("AUGUR_RECURSION_LIMIT")
 if recursion_limit:
@@ -40,7 +39,7 @@ command_strings = [
     "export",
     "validate",
     "version",
-    "import",
+    "import_",
     "measurements",
 ]
 
@@ -51,26 +50,11 @@ def make_parser():
         prog        = "augur",
         description = "Augur: A bioinformatics toolkit for phylogenetic analysis.")
 
-    subparsers = parser.add_subparsers()
-
     add_default_command(parser)
     add_version_alias(parser)
 
-    for command in COMMANDS:
-        # Add a subparser for each command.
-        subparser = subparsers.add_parser(
-            command_name(command),
-            help        = first_line(command.__doc__),
-            description = command.__doc__)
-
-        subparser.set_defaults(__command__ = command)
-
-        # Let the command register arguments on its subparser.
-        command.register_arguments(subparser)
-
-        # Use the same formatting class for every command for consistency.
-        # Set here to avoid repeating it in every command's register_parser().
-        subparser.formatter_class = argparse.ArgumentDefaultsHelpFormatter
+    subparsers = parser.add_subparsers()
+    add_command_subparsers(subparsers, COMMANDS)
 
     return parser
 
@@ -99,18 +83,6 @@ def run(argv):
         sys.exit(2)
 
 
-def add_default_command(parser):
-    """
-    Sets the default command to run when none is provided.
-    """
-    class default_command():
-        def run(args):
-            parser.print_help()
-            return 2
-
-    parser.set_defaults(__command__ = default_command)
-
-
 def add_version_alias(parser):
     """
     Add --version as a (hidden) alias for the version command.
@@ -129,17 +101,3 @@ def add_version_alias(parser):
         nargs  = 0,
         help   = argparse.SUPPRESS,
         action = run_version_command)
-
-
-def command_name(command):
-    """
-    Returns a short name for a command module.
-    """
-
-    def remove_prefix(prefix, string):
-        return re.sub('^' + re.escape(prefix), '', string)
-
-    package     = command.__package__
-    module_name = command.__name__
-
-    return remove_prefix(package, module_name).lstrip(".").replace("_", "-")