+++ augur/parse.py
@@ -137,7 +137,8 @@ def parse_sequence(sequence, fields, strain_key="strain", separator="|", prettif
     return sequence, metadata
 
 
-def register_arguments(parser):
+def register_parser(parent_subparsers):
+    parser = parent_subparsers.add_parser("parse", help=__doc__)
     parser.add_argument('--sequences', '-s', required=True, help="sequences in fasta or VCF format")
     parser.add_argument('--output-sequences', help="output sequences file")
     parser.add_argument('--output-metadata', help="output metadata file")
@@ -146,6 +147,7 @@ def register_arguments(parser):
     parser.add_argument('--separator', default='|', help="separator of fasta header")
     parser.add_argument('--fix-dates', choices=['dayfirst', 'monthfirst'],
                                 help="attempt to parse non-standard dates and output them in standard YYYY-MM-DD format")
+    return parser
 
 
 def run(args):